requirements (2).txt
streamlit
PyPDF2
langchain
langchain-openai
langchain-community
faiss-cpu
openai
tiktoken
python-dotenv
langchain-classic
